+++ mtgogui/src/menubar.rs
@@ -6,16 +6,18 @@ use fltk::{
     menu::{self, MenuFlag},
     misc::Progress,
     prelude::*,
-    text::{self, TextAttr, TextBuffer, TextDisplay, WrapMode},
+    text::{TextAttr, TextBuffer, TextDisplay, WrapMode},
     window::Window,
 };
 use fltk_flex::Flex;
 use mtgoupdater::{
     mtgo_preprocessor_api::run_mtgo_preprocessor_version, mtgogetter_api::mtgogetter_version,
 };
-use std::path::PathBuf;
+use std::{io, path::PathBuf};
 use util::ProgressUpdate;
 
+mod about;
+mod setup;
 pub mod util;
 
 /// Messages that can be received by the menubar
@@ -29,16 +31,16 @@ pub enum MenubarMessage {
 }
 
 /// The menubar for the application
-pub(super) struct McmMenuBar {
-    pub(super) menu: menu::SysMenuBar,
+pub struct McmMenuBar {
+    menu: menu::SysMenuBar,
     ev_emitter: app::Sender<Message>,
-    pub(super) progress_bar: Progress,
+    progress_bar: Progress,
 }
 
 impl McmMenuBar {
     pub fn new(w: i32, h: i32, s: &app::Sender<Message>) -> Self {
         let mut mb = menu::SysMenuBar::default().with_size(w, h);
-        init_menu_bar(&mut mb, s);
+        setup::init_menu_bar(&mut mb, s);
 
         let progress_bar_width = 300;
         let mut progress = Progress::new(
@@ -67,7 +69,7 @@ impl McmMenuBar {
         match ev {
             MenubarMessage::Open => self.open_full_tradelist(),
             MenubarMessage::Quit => app::quit(),
-            MenubarMessage::About => show_about(),
+            MenubarMessage::About => about::show_about(),
             MenubarMessage::Example => todo!("example"),
             MenubarMessage::ProgressBar(update) => {
                 if update.show {
@@ -97,171 +99,5 @@ impl McmMenuBar {
                 dialog::alert(center().0 - 200, center().1 - 100, "File does not exist!")
             }
         }
-    }
-}
-
-/// Get version information from all components and display in a pop-up window
-pub(super) fn show_about() {
-    // Start by getting the versions from MTGO Getter and MTGO Preprocessor
-    let mtgogetter_version = match mtgogetter_version() {
-        Ok(v) => v,
-        Err(e) => {
-            log::info!("Error getting mtgogetter version: {e}");
-            dialog::alert(
-                center().0 - 200,
-                center().1 - 100,
-                &format!("Could not find MTGO Getter binary!\n{e}"),
-            );
-            return;
-        }
-    };
-    let mtgo_preproc_version = match run_mtgo_preprocessor_version() {
-        Ok(v) => v,
-        Err(e) => {
-            log::info!("Error getting mtgo preprocessor version: {e}");
-            dialog::alert(
-                center().0 - 200,
-                center().1 - 100,
-                &format!("Could not find MTGO Preprocessor binary!\n{e}"),
-            );
-            return;
-        }
-    };
-
-    let mtgogetter_version_str = String::from_utf8_lossy(&mtgogetter_version.stdout);
-    let version_pos = mtgogetter_version_str.trim().find("version ").unwrap();
-    let just_v = &mtgogetter_version_str[version_pos + 8..];
-    let preproc_version_str = String::from_utf8_lossy(&mtgo_preproc_version.stdout)
-        .trim()
-        .to_string();
-
-    let mtgoupdater_version = mtgoupdater::mtgo_updater_version();
-
-    let mtgo_gui_version = env!("CARGO_PKG_VERSION");
-
-    let project_url = "https://github.com/CramBL/mtgo-collection-manager/";
-
-    let w_width = 450;
-    let mut tbuf = TextBuffer::default();
-    let mut sbuf = TextBuffer::default();
-    let mtgo_cm_txt = format!(
-        "{:^width$}\n",
-        "MTGO Collection Manager",
-        width = 65 - "MTGO Collection Manager".len()
-    );
-    let mtgo_cm_ver_txt = format!(
-        "{:^width$}\n\n",
-        format!("v{}", mtgo_gui_version),
-        width = 57 - format!("v{}", mtgo_gui_version).len()
-    );
-    tbuf.set_text(&mtgo_cm_txt);
-    sbuf.set_text(&"A".repeat(mtgo_cm_txt.len()));
-    tbuf.append(&mtgo_cm_ver_txt);
-    sbuf.append(&"A".repeat(mtgo_cm_ver_txt.len()));
-    tbuf.append("Components:\n");
-    sbuf.append(&"B".repeat("Components:\n".len()));
-    let component_left_pad = 20;
-    let mtgogetter_txt = format!(
-        "   {:<width$} v{}",
-        "MTGO Getter",
-        just_v,
-        width = component_left_pad
-    );
-    tbuf.append(&mtgogetter_txt);
-    sbuf.append(&"C".repeat(mtgogetter_txt.len()));
-    let mtgopreproc_txt = format!(
-        "   {:<width$} {}\n",
-        "MTGO Preprocessor",
-        preproc_version_str,
-        width = component_left_pad
-    );
-    tbuf.append(&mtgopreproc_txt);
-    sbuf.append(&"C".repeat(mtgopreproc_txt.len()));
-    let mtgoupdater_txt = format!(
-        "   {:<width$} v{}\n\n",
-        "MTGO Updater",
-        mtgoupdater_version,
-        width = component_left_pad
-    );
-    tbuf.append(&mtgoupdater_txt);
-    sbuf.append(&"C".repeat(mtgoupdater_txt.len()));
-    tbuf.append("Homepage:\n");
-    sbuf.append(&"B".repeat("Homepage:\n".len()));
-    tbuf.append(project_url);
-    sbuf.append(&"D".repeat(project_url.len()));
-
-    let txt_lines = tbuf.count_lines(0, tbuf.length());
-
-    let mut win = Window::default()
-        .with_size(w_width, txt_lines * 30)
-        .with_pos(center().0 - 300, center().1 - 100)
-        .with_label(&format!(
-            "About MTGO Collection Manager v{}",
-            mtgo_gui_version
-        ));
-    win.set_icon(Some(assets::get_logo()));
-    let flex_about = Flex::default()
-        .with_pos(0, 0)
-        .with_align(Align::Center)
-        .size_of_parent()
-        .column();
-
-    let mut txt_disp = TextDisplay::default();
-    txt_disp.align();
-    txt_disp.set_buffer(tbuf);
-    txt_disp.set_highlight_data_ext(sbuf, util::TEXT_ABOUT_STYLES);
-    txt_disp.set_align(Align::Center);
-    txt_disp.wrap_mode(WrapMode::AtBounds, 0);
-    txt_disp.set_text_color(Color::White);
-
-    flex_about.end();
-    win.end();
-    win.show();
-}
-
-fn init_menu_bar(menu: &mut menu::SysMenuBar, s: &Sender<Message>) {
-    menu.set_frame(FrameType::FlatBox);
-
-    menu.add_emit(
-        "&File/Open Full Trade list...\t",
-        Shortcut::Ctrl | 'o',
-        MenuFlag::Normal,
-        s.clone(),
-        MenubarMessage::Open.into(),
-    );
-
-    menu.add_emit(
-        "&File/Quit\t",
-        Shortcut::Ctrl | 'q',
-        MenuFlag::Normal,
-        s.clone(),
-        Message::Quit,
-    );
-
-    menu.add_emit(
-        "&Help/About\t",
-        Shortcut::None,
-        MenuFlag::Normal,
-        s.clone(),
-        MenubarMessage::About.into(),
-    );
-
-    if let Some(mut item) = menu.find_item("&File/Quit\t") {
-        item.set_label_color(Color::Red);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_init_menu_bar() {
-        let mut menu = menu::SysMenuBar::default();
-        let (s, _) = app::channel();
-        init_menu_bar(&mut menu, &s);
-        assert!(menu.find_item("&File/Open Full Trade list...\t").is_some());
-        assert!(menu.find_item("&File/Quit\t").is_some());
-        assert!(menu.find_item("&Help/About\t").is_some());
     }
 }
